vector_db/qdrant/requirements.txt
boto3==1.28.84
numpy==1.24.4
openai==1.2.0
pandas==1.5.3
python-dotenv==1.0.1
qdrant_client==1.7.2
sentence_transformers==2.2.2
torch==2.0.1
tqdm==4.65.0